requirements.txt
# ===========================
# Core Dependencies
# ===========================
streamlit>=1.28.0

# ===========================
# LangChain Ecosystem
# ===========================
langchain>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
langchain-ollama>=0.2.0
langchain-chroma>=0.1.4
langgraph>=0.2.0
langchain-text-splitters>=0.3.0

# ===========================
# Vector Store
# ===========================
chromadb>=0.4.0

# ===========================
# Environment & Utils
# ===========================
python-dotenv>=1.0.0

# ===========================
# HTTP Client (for API calls)
# ===========================
httpx>=0.25.0
